examples/requirements.txt
# Python MySQL 驱动依赖
mysql-connector-python>=8.0.0
pymysql>=1.0.0
